pyproject.toml
[project]
name = "bot-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "py-cord>=2.6.1",
    "python-dotenv>=1.1.1",
]
